+++ gpxviewer/ui_plotwindow.py
@@ -35,12 +35,18 @@ class Ui_PlotWindow(object):
         self.actExportSelectedSize = QtWidgets.QAction(PlotWindow)
         self.actExportSelectedSize.setIcon(icon)
         self.actExportSelectedSize.setObjectName("actExportSelectedSize")
+        self.actFitWidth = QtWidgets.QAction(PlotWindow)
+        self.actFitWidth.setShortcut("Ctrl+R")
+        self.actFitWidth.setObjectName("actFitWidth")
         self.toolBar.addAction(self.actExportCurrentSize)
         self.toolBar.addAction(self.actExportSelectedSize)
+        self.toolBar.addSeparator()
+        self.toolBar.addAction(self.actFitWidth)
 
         self.retranslateUi(PlotWindow)
         self.actExportCurrentSize.triggered.connect(PlotWindow.onSaveCurrentSize)
         self.actExportSelectedSize.triggered.connect(PlotWindow.onSaveSelectedSize)
+        self.actFitWidth.triggered.connect(PlotWindow.onFitWidth)
         QtCore.QMetaObject.connectSlotsByName(PlotWindow)
 
     def retranslateUi(self, PlotWindow):
@@ -49,4 +55,5 @@ class Ui_PlotWindow(object):
         self.actExportCurrentSize.setToolTip(_translate("PlotWindow", "Export current size"))
         self.actExportSelectedSize.setText(_translate("PlotWindow", "Export selected size"))
         self.actExportSelectedSize.setToolTip(_translate("PlotWindow", "Export selected size"))
+        self.actFitWidth.setText(_translate("PlotWindow", "Fit width"))
 from gpxviewer.plotcanvas import PlotCanvas